app/requirements.txt
Django==5.0
djangorestframework==3.14.0
gunicorn==21.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.3.1